+++ packages/picimo/src/components/TextureRef.tsx
@@ -1,14 +1,14 @@
 import '@react-three/fiber';
 import {ForwardedRef, forwardRef} from 'react';
 import {Texture} from 'three';
-import {useTextureRef} from '../hooks/useTextureRef';
+import {useTexture} from '../hooks/useTexture';
 
 export type TextureRefProps = JSX.IntrinsicElements['texture'] & {
   name: string | symbol;
 };
 
 function Component({name, children, ...props}: TextureRefProps, ref: ForwardedRef<Texture>) {
-  const texture = useTextureRef(name);
+  const texture = useTexture(name);
 
   return texture ? (
     <primitive object={texture} ref={ref} {...props}>